use utf8 option for default tester codec
I had mistakenly(?) assumed that HTTP::Response's decoded_content
would decode JSON from utf8.  I still think it used to do this, but I
didn't do any deep diving to check.  It doesn't now, and maybe it never
did.  Because I thought that, though, we assumed that ->decoded_content
was returning text, then decoded the JSON as if it was a character
string, which meant we got mojibake.  The solution is to simplify!