expand how authentication can be used to configure jmap tester
* jmap_uri is now called api_uri, to match the apiUrl from the auth token
  (and those other URIs point to JMAP things, too!)

* all URIs are now read-write and optional; we barf if they are unset but
  needed, at runtime

* update_uris_from_auth sets all the URI attributes based on what we're told
  by the auth struct

Note that right now we won't clear URIs if the auth struct has nothing, which
means that using a tester for two different authentications might have weird
results, if the second auth does not set a URI that the first did.

We also do not (yet) set a default Authorization header when accepting an
auth object.