abstract out the UA role and allow futures
Okay, so probably this could be two commits, and maybe we can rewrite
this commit later, but bear with me for now.

The end goal here is to allow the JMAP client to work in an async
program, so that ->request can return a future that will resolve to the
sentence collection when it's ready.  Meanwhile, the changes should not
*seriously* break existing code.  So, it's okay if a small amount of
mechanical change is needed for code that was calling things directly on
the LWP object, but we shouldn't break every call to ->request in
existing tests.

We accomplish this through a few steps.

First, we abstract out $tester->ua, previously a LWP::UserAgent object,
into a thin adapter over the client of your choice.  LWP remains the
default, but I've added a stub implementation for Net::Async::HTTP,
which should work just fine, but will break logging.  (A few things in
this commit worsen logging, some fatally for now.)

Next, we define the UA adapters to return futures, not responses.  The
JMAP::Tester code that interprets those responses is then moved into
a ->then sequence that hangs off the HTTP response.  Any user-facing
method that works this way will either return a future or a Result,
depending on the JMAP::Tester's return_futures setting.  The default is
to return results, to match existing behavior.

If futures are returned, JMAP::Tester::Result::Failure objects show up
as failures of the future.  If results are returned, Failure results may
be returned, but other forms of failing futures will result in
exceptions.