work to get auto-diagnostics on aborts
We need to twist the object graph around a bit so that when an abort is
being generated, we have access to the whole response.

Otherwise, an individual sentence is unable to throw an abort that shows
the enclosing response, which is probably what we want most of the time.