provide a pluggable diagnostics generator
* when no diagnostics are provided, provide some!
* let the provider be pluggable; starts with JSON, later Data::Printer?

The current pluggability of the diagnostician is not great.  I think it
should probably be something like:

1. you can replace the pretty printer from JSON to something else
2. you can, possibly per-abort, describe what diagnostics to generate

So maybe:

    $res = Result->new({
      ...,
      diagnostic_dumper => sub { ... },
      default_diagnostics => [
        'Response sentences' => sub ($res) { [ $res->items ] },
        ...
      ],
    });

Then diagnostics can be requested with:

    ->abort($message, [
      $label => $value,
      ...
    ]);

The second argument is a optlist (q.v. Data::OptList).  If there is no
value, the label is just a line.  If there is a value...

  ...if a subref, it's called with the $res as an argument.  The result
  is diagnostically dumpered.

  ...if another kind of ref, it's diagnostically dumpered.

We use OptList in part so that order is specified and not merely
cmp-based.