add JSONLiteral, so you can test malformed or just weird JSON
Say there's a bug in your handling of space around the colon in a JSON
object, or some other nonsense.  Now you can...

  use JMAP::Tester::Sugar qw(json_literal);

  my $res = $tester->request( json_literal('{"methodCalls"  :  [...') );

This will bypass all of the processing of the input to the request.  It
will be taken as the *exact* JSON to send.  A wrapper won't be created,
call ids won't be added, and so on.  It won't even be JSON encoded,
it'll just be passed in.